+++ src/rics/_internal_support/__init__.py
@@ -1,3 +1,0 @@
-import rics._internal_support._local_or_remote as local_or_remote
-
-__all__ = ["local_or_remote"]