+++ src/rics/utility/__init__.py
@@ -1,5 +1,0 @@
-"""Utility modules."""
-
-from rics.utility._just_the_way_i_like_it import configure_stuff
-
-__all__ = ["configure_stuff"]